Refactoring for the single file upload and PDF conversions

+++ src/helpers/fileRemover.ts
@@ -1,24 +1,20 @@
 import { winstonLogger } from '../util/winstonLogger';
 
-const fs = require("fs");
-const path = require("path");
+const fs = require('fs');
+const path = require('path');
 
 export function rmDir(dirPath, removeSelf) {
-    try {
-        const files = fs.readdirSync(dirPath);
-        if (files.length > 0)
-        for (let i = 0; i < files.length; i++) {
-            const filePath = path.join(dirPath, files[i]);
-            if (fs.statSync(filePath).isFile())
-                fs.unlinkSync(filePath);
-            else
-                rmDir(filePath, true);
-        }
-        if (removeSelf)
-        fs.rmdirSync(dirPath);
-        }
-        catch (e) {
-            winstonLogger.error(e);
-            return;
-        }
+  try {
+    const files = fs.readdirSync(dirPath);
+    if (files.length > 0)
+      for (let i = 0; i < files.length; i++) {
+        const filePath = path.join(dirPath, files[i]);
+        if (fs.statSync(filePath).isFile()) fs.unlinkSync(filePath);
+        else rmDir(filePath, true);
+      }
+    if (removeSelf) fs.rmdirSync(dirPath);
+  } catch (e) {
+    winstonLogger.error(e);
+    return;
   }
+}
